cli: Add `tracker extract` command
This wraps `$libexecdir/tracker-extract --file`. I think that the
extractor is a really important part of Tracker that needs to be
accessible in a standard way.

Currently the output of `tracker-extract --file` contains a lot of
log information even with `--verbosity=0`. And dumping the result as
a SPARQL UPDATE operation isn't ideal, it'd be nicer to output Turtle
or something.

https://bugzilla.gnome.org/show_bug.cgi?id=751991